server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.2
    # via pyjwt
dj-rest-auth[with-social,with_social]==7.0.1
    # via -r requirements.in
django==5.2
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth[socialaccount]==65.7.0
    # via dj-rest-auth
django-environ==0.12.0
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
idna==3.10
    # via requests
markdown==3.8
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.9.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
requests==2.32.3
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
sqlparse==0.5.3
    # via django
urllib3==2.4.0
    # via requests
